Updated all js into one file dynamically linked to flask; instituted limit of 500 for flask server.
