+++ flask_app.py
@@ -8,6 +8,8 @@ from sqlalchemy.ext.automap import automap_base
 from sqlalchemy.orm import Session
 from sqlalchemy import create_engine
 from sqlalchemy import func
+from sqlalchemy import cast
+from sqlalchemy import String
 import json
 
 # # Define the database connection parameters
@@ -44,14 +46,14 @@ def DictionaryRoute():
     
     # Query database
     session = Session(engine)
-    table_data = session.query(table.inspection_id, table.dba_name, table.facility_type, table.risk, table.address, table.inspection_date, table.inspection_type, table.results, table.violations, table.location).all()
+    table_data = session.query(table.inspection_id, table.dba_name, table.facility_type, table.risk, table.address, table.inspection_date, table.inspection_type, table.results, table.violations, table.latitude, table.longitude, table.location).all()
     session.close()
     
     # Create a list of dictionaries
 
     all_inspections = []
 
-    for inspection_id, dba_name, facility_type, risk, address, inspection_date, inspection_type, results, violations, location in table_data:
+    for inspection_id, dba_name, facility_type, risk, address, inspection_date, inspection_type, results, violations, latitude, longitude, location in table_data:
         dict = {}
         dict["inspection_id"] = inspection_id
         dict["dba_name"] = dba_name
@@ -62,37 +64,40 @@ def DictionaryRoute():
         dict["inspection_type"] = inspection_type
         dict["results"] = results
         dict["violations"] = violations
+        dict["latitude"] = latitude
+        dict["longitude"] = longitude
         dict["location"] = location
         all_inspections.append(dict)
 
     # Return results
     return jsonify(all_inspections)
 
-@app.route("/api/<vID>/<vName>/<vType>/<vRisk>/<vResults>/<vViolations>")
+@app.route("/api/<vID>/<vName>/<vType>/<vRisk>/<vResults>")
 
 # Filter examples:
 
-# /api/CHIPOTLE/%/%/%/%   =   returns all inspections for places named "Chipotle"
-# /api/chip/restaurant/3/pass/%   =   returns all places containing "chip" where the type contains "restaurant" and the risk level contains "3" and the results includes "pass"
-# /api/%/restaurant/%/fail/food   =   returns all places where the type contains "restaurant" and the results includes "fail" and the violations include "food"
+# /api/%/CHIPOTLE/%/%/%  =   returns all inspections for places named "Chipotle"
+# /api/%/chip/restaurant/3/pass   =   returns all places containing "chip" where the type contains "restaurant" and the risk level contains "3" and the results includes "pass"
+# /api/%/%/restaurant/%/fail   =   returns all places where the type contains "restaurant" and the results includes "fail"
 
-def FilteredRoute(vID="%",vName="%",vType="%",vRisk="%",vResults="%",vViolations="%"):
+def FilteredRoute(vID="%",vName="%",vType="%",vRisk="%",vResults="%"):
 
     # Query database
     session = Session(engine)
-    table_data = session.query(table.inspection_id, table.dba_name, table.facility_type, table.risk, table.address, table.inspection_date, table.inspection_type, table.results, table.violations, table.location)\
+    table_data = session.query(table.inspection_id, table.dba_name, table.facility_type, table.risk, table.address, table.inspection_date, table.inspection_type, table.results, table.violations, table.latitude, table.longitude, table.location)\
+        .filter(cast (table.inspection_id, String).contains(vID))\
         .filter(func.upper(table.dba_name).contains(func.upper(vName)))\
         .filter(func.upper(table.facility_type).contains(func.upper(vType)))\
         .filter(func.upper(table.risk).contains(func.upper(vRisk)))\
         .filter(func.upper(table.results).contains(func.upper(vResults)))\
-        .filter(func.upper(table.violations).contains(func.upper(vViolations)))
+        .limit(500).all()
     session.close()
 
     # Create a list of dictionaries
 
     filtered_inspections = []
 
-    for inspection_id, dba_name, facility_type, risk, address, inspection_date, inspection_type, results, violations, location in table_data:
+    for inspection_id, dba_name, facility_type, risk, address, inspection_date, inspection_type, results, violations, latitude, longitude, location in table_data:
         dict = {}
         dict["inspection_id"] = inspection_id
         dict["dba_name"] = dba_name
@@ -103,6 +108,8 @@ def FilteredRoute(vID="%",vName="%",vType="%",vRisk="%",vResults="%",vViolations
         dict["inspection_type"] = inspection_type
         dict["results"] = results
         dict["violations"] = violations
+        dict["latitude"] = latitude
+        dict["longitude"] = longitude
         dict["location"] = location
         filtered_inspections.append(dict)
