+++ static/css/style.css
@@ -13,6 +13,11 @@ body
   font-family: 'Helvetica Neue', Arial, sans-serif;
 }
 
+#table-area {
+  height: 400px;
+  overflow: scroll;
+}
+
 
 .legend {
   padding: 10px;
