+++ templates/index.html
@@ -47,7 +47,7 @@
           <div class="panel panel-default">
             <div class="panel-heading">Search</div>
             <div class="panel-body">
-              <form>
+              <form id="mainForm">
                <div class="form-group">
                 <ul class="list-group">
                   <li class="filter list-group-item">
@@ -123,7 +123,7 @@
           <div class="caption">
             <h3>Facility Type Distribution</h3>
             <!-- The div where we will inject our facility distribution pie chart -->
-            <div class="card">           
+            <div id= "donut" class="card">           
               <!-- Custom tooltip canvas -->
               <canvas id="tooltip-canvas"></canvas>
               <!-- Canvas for Chart.js -->
@@ -163,13 +163,8 @@
   <script src="https://cdn.polyfill.io/v2/polyfill.min.js"></script>
 
   
-  <!-- Map -->
-  <script src="../static/js/logic.js"></script>
   <!-- Table -->
   <script src="../static/js/main.js"></script>
-  <!-- Pie Chart-->
-  <script src="../static/js/piechart.js"></script> 
-
 
 
 </body>
